[chore] migrate the rest of integration_test.go to perms_store_test.go
